Agrego medicion de bateria - hay que chequear con voltimetro

+++ main_arduino/menu.h
@@ -239,7 +239,7 @@ void menu_pacientes_fields();
  * @param <refresh> <Refresca la pantalla>
  *
  */
-void batteryStatus(uint8_t x);
+void batteryStatus(float x);
 
 /**
  * @brief <Imprime el menu principal en el display>
